Roles with modifications in Game and User Controller´s

+++ diceGame/app/Http/Controllers/GameController.php
@@ -11,48 +11,71 @@ class GameController extends Controller
 
     public function roll($id)
     {
-        $user = User::find($id);
-        //totla game
-        //total wins 
-        
-        $dice = rand(1,5) + rand(1,5);
+        $user = auth()->user();
 
-        if($dice == 7){
+        if($user->admin_roll == 'Admin' || $user->id == $id)
+        {
 
-            $user->total_wins = $user->total_wins + 1;
-        
+                $user = User::find($id);
+                //totla game
+                //total wins 
+                
+                $dice = rand(1,5) + rand(1,5);
+
+                if($dice == 7){
+
+                    $user->total_wins = $user->total_wins + 1;
+                
+                }
+
+                $user->total_games = $user->total_games + 1;
+
+                $user->winning_percentage = ($user->total_wins * 100) / $user->total_games;
+
+                $user->save();
+
+                
+                $game = Game::create([
+                    
+                    'player_id' => $id,
+                    'result' => $dice,
+                    
+                    
+                ]);
+
+            return  $game;
+
+        }else{
+
+            return response([
+                'error_message' => "Sorry! You don't have access"
+            ], 401 );
+
         }
-
-        $user->total_games = $user->total_games + 1;
-
-        $user->winning_percentage = ($user->total_wins * 100) / $user->total_games;
-
-        $user->save();
-
-        
-        $game = Game::create([
-            
-            'player_id' => $id,
-            'result' => $dice,
-            
-            
-        ]);
-
-       return  $game;
     }
 
 
     public function destroy($id)
     {
+        $user = auth()->user();
 
-        foreach(Game::all() as $game){
+        if($user->admin_roll == 'Admin' || $user->id == $id)
+        {
 
-            if($game->player_id == $id){
+            foreach(Game::all() as $game){
 
-                Game::destroy($game->id);
-     
+                if($game->player_id == $id){
+
+                    Game::destroy($game->id);
+        
+                }
             }
 
+        }else{
+
+            return response([
+                'error_message' => "Sorry! You don't have access"
+            ], 401 );
 
         }
     
@@ -72,7 +95,7 @@ class GameController extends Controller
     public function rank_loser()
     {
 
-        $player = User::orderBy('winning_percentage', 'asc')->get()->first();
+        $player = User::where('total_games', '!=', null)->orderBy('winning_percentage', 'asc')->get()->first();
       
         return $player;
